feat(rating): add is librarian to delete rating

+++ src/main/java/fr/eni/bookhubbackend/service/RatingService.java
@@ -2,6 +2,7 @@ package fr.eni.bookhubbackend.service;
 
 import fr.eni.bookhubbackend.entity.bo.Rating;
 import fr.eni.bookhubbackend.entity.bo.User;
+import fr.eni.bookhubbackend.entity.enums.RoleEnum;
 import fr.eni.bookhubbackend.repository.RatingRepository;
 import fr.eni.bookhubbackend.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -32,7 +33,15 @@ public class RatingService {
         ratingRepository.save(rating);
     }
 
-    public void deleteRating(final Long id) {
-         ratingRepository.deleteById(id);
+    public void deleteRating(final Long id, String email) {
+        boolean isLibrarian = userRepository.findByEmail(email)
+                .map(user -> user.getRole() == RoleEnum.LIBRARIAN || user.getRole() == RoleEnum.ADMIN)
+                .orElse(false);
+
+        if (!isLibrarian) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not allowed to delete this rating");
+        }
+
+        ratingRepository.deleteById(id);
     }
 }
